requirements.txt
# Core Framework
streamlit>=1.31.0

# LLM & Embeddings
openai>=1.12.0

# Vector Database
chromadb>=0.4.22

# PDF Processing
PyMuPDF>=1.23.0

# Database
SQLAlchemy>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Utilities
requests>=2.31.0
